sce-models-api/sce-models-api-develop/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
flasgger==0.9.7.1
opencv-python==4.11.0.86
Pillow==11.3.0
NumPy==2.1.0
tensorflow==2.19.0
pika==1.3.2
python-dotenv==1.1.0
psycopg2-binary==2.9.10  #para prod
paramiko==3.5.1  #para prod
scikit-learn==1.6.1
# Modelo Binario
torch==2.8.0+cpu
torchvision==0.23.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

safetensors==0.6.2
transformers==4.57.3
tokenizers==0.22.1
rfdetr==1.3.0
pdf2image==1.17.0